agent/src/pyproject.toml
[project]
name = "edwardo"
version = "0.1.0"
description = "Autonomous life assistant with multi-agent ADK orchestration and MCP integrations"
readme = "README.md"
authors = [
    { name = "mohammed-alaa40123", email = "mohammedalaa40123@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=0.3.1",
    "mcp>=1.21.2",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "pymongo>=4.11.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]

[project.scripts]
edwardo = "edwardo:main"

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"
